+++ CHANGELOG.md
@@ -31,6 +31,7 @@
  - Implement test service for w3c/distributed-tracing test harness (#293)
  - End HTTP client spans on response body closure (#289)
  - module/apmgrpc requires Go 1.9+ (#300)
+ - Invalid tag key characters are replaced with underscores (#308)
 
 ## [v0.5.2](https://github.com/elastic/apm-agent-go/releases/tag/v0.5.2)